added extra step delete profile

+++ src/app/user-profile/user-profile.component.ts
@@ -80,26 +80,25 @@ export class UserProfileComponent implements OnInit {
   }
 
   deleteUser(): void {
-    this.fetchApiData.deleteUser().subscribe(
-      (resp: any) => {
-        this.snackBar.open(
-          'Your account has been deleted!',
-          'Ok',
-          { 
-            duration: 2000,
-          }
-        );
+    let check = confirm(
+      'This will delete your account!'
+    );
+    if (check) {
+      this.fetchApiData.deleteUser().subscribe(() => {
         localStorage.clear();
-      },
-      (result) => {
-        this.snackBar.open(result, 'OK', {
+        this.router.navigate(['welcome']);
+        this.snackBar.open('Account has been deleted', 'OK', {
           duration: 2000,
         });
-        this.router.navigate(['/welcome']).then(() => {
-          window.location.reload();
-        });
-      }
-    );
+      });
+    } else {
+      window.location.reload();
+    }
+  }
+  profileUpdateDialog(): void {
+    this.dialog.open(UserProfileUpdateComponent, {
+      panelClass: 'update-dialog',
+    });
   }
 
 }
